Ensure that 'start logging' displays correct filename
Previously the filename was calculated twice: first when opening the
file, and then when displaying the message.  There is a small chance
that the time would change between those two events, causing the
displayed filename to be incorrect.  This change calculates the name
once and then uses that name in both places.